requirements.txt
crewai==0.130.0
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
python-multipart==0.0.20
pypdf==5.6.0
pydantic>=2.0,<3.0
litellm>=1.30.0
sqlalchemy==2.0.41
httpx>=0.27.0
